Create convertNiFiSeperatorToSpace
Using ExecuteScript processor in NiFi has a default separator as semi-colon(;) to separate the arguments. While using a Unix script call from NiFi processor with arguments, the semicolon isn't translated properly to Unix compatible space character.
The idea of this script is to push all the arguments (as-is) from NiFi environment to Unix environment and then convert semi-colons to space.